move credentials.py
