+++ lib/streamlit/runtime/app_session.py
@@ -22,7 +22,6 @@ from typing import TYPE_CHECKING, Callable, Dict, Optional, List, Union
 import streamlit.elements.exception as exception_utils
 from streamlit import __version__, config, source_util, secrets
 from streamlit.case_converters import to_snake_case
-from streamlit.credentials import Credentials
 from streamlit.logger import get_logger
 from streamlit.metrics_util import Installation
 from streamlit.proto.BackMsg_pb2 import BackMsg
@@ -38,6 +37,7 @@ from streamlit.proto.NewSession_pb2 import (
 from streamlit.proto.PagesChanged_pb2 import PagesChanged
 from streamlit.watcher import LocalSourcesWatcher
 from . import caching, legacy_caching
+from .credentials import Credentials
 from .in_memory_file_manager import in_memory_file_manager
 from .scriptrunner import (
     RerunData,
